# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export type BuildState = typeof BuildState[keyof typeof BuildState];


export const BuildState = {
  IDLE: 'IDLE',
  UPLOADING: 'UPLOADING',
  CONNECTING_TO_TERMUX: 'CONNECTING_TO_TERMUX',
  ANALYZING: 'ANALYZING',
  BUILDING: 'BUILDING',
  VALIDATING_APK: 'VALIDATING_APK',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
} as const;

export interface ApiError {
  error: string;
  code: string;
  state: BuildState;
  diagnostics?: string[];
}

export type BuilderConnectionStatus = typeof BuilderConnectionStatus[keyof typeof BuilderConnectionStatus];


export const BuilderConnectionStatus = {
  connected: 'connected',
} as const;

export type BuilderConnectionToolsItem = typeof BuilderConnectionToolsItem[keyof typeof BuilderConnectionToolsItem];


export const BuilderConnectionToolsItem = {
  health_check: 'health_check',
  get_environment: 'get_environment',
  list_apk_outputs: 'list_apk_outputs',
  build_project: 'build_project',
} as const;

export type BuilderConnectionEnvironment = { [key: string]: unknown };

export interface BuilderConnection {
  connected: boolean;
  status: BuilderConnectionStatus;
  tools: BuilderConnectionToolsItem[];
  checkedAt: string;
  environment?: BuilderConnectionEnvironment;
}

export interface AndroidBuildInput {
  /**
     * @minLength 5
     * @maxLength 512
     * @pattern ^/(?:storage/emulated/0|sdcard)/(?!.*(?:^|/)\.\.(?:/|$))[A-Za-z0-9 _.,()@+\-/]+\.zip$
     */
  androidZipPath: string;
  developerMode?: boolean;
}

export interface ApkArtifact {
  filename: string;
  phonePath?: string;
  /** @nullable */
  sizeBytes?: number | null;
  generatedAt: string;
  status: string;
  /** @nullable */
  sha256?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}

export type BuildTimelineItemStatus = typeof BuildTimelineItemStatus[keyof typeof BuildTimelineItemStatus];


export const BuildTimelineItemStatus = {
  complete: 'complete',
  active: 'active',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface BuildTimelineItem {
  state: BuildState;
  label: string;
  status: BuildTimelineItemStatus;
}

export interface BuildResult {
  state: BuildState;
  message: string;
  artifact: ApkArtifact;
  timeline: BuildTimelineItem[];
  logs: string[];
}