# lib/api-zod/src/generated/types/androidBuildInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */

export interface AndroidBuildInput {
  /**
     * @minLength 5
     * @maxLength 512
     * @pattern ^/(?:storage/emulated/0|sdcard)/(?!.*(?:^|/)\.\.(?:/|$))[A-Za-z0-9 _.,()@+\-/]+\.zip$
     */
  androidZipPath: string;
  developerMode?: boolean;
}