# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */
import type { BuildState } from './buildState';

export interface ApiError {
  error: string;
  code: string;
  state: BuildState;
  diagnostics?: string[];
}