# lib/api-zod/src/generated/types/apkArtifact.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */

export interface ApkArtifact {
  filename: string;
  phonePath?: string;
  /** @nullable */
  sizeBytes?: number | null;
  generatedAt: Date;
  status: string;
  /** @nullable */
  sha256?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}