# lib/api-zod/src/generated/types/buildResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */
import type { ApkArtifact } from './apkArtifact';
import type { BuildState } from './buildState';
import type { BuildTimelineItem } from './buildTimelineItem';

export interface BuildResult {
  state: BuildState;
  message: string;
  artifact: ApkArtifact;
  timeline: BuildTimelineItem[];
  logs: string[];
}