# lib/api-zod/src/generated/types/buildState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */

export type BuildState = typeof BuildState[keyof typeof BuildState];


export const BuildState = {
  IDLE: 'IDLE',
  UPLOADING: 'UPLOADING',
  CONNECTING_TO_TERMUX: 'CONNECTING_TO_TERMUX',
  ANALYZING: 'ANALYZING',
  BUILDING: 'BUILDING',
  VALIDATING_APK: 'VALIDATING_APK',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
} as const;