# lib/api-zod/src/generated/types/buildTimelineItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */
import type { BuildState } from './buildState';
import type { BuildTimelineItemStatus } from './buildTimelineItemStatus';

export interface BuildTimelineItem {
  state: BuildState;
  label: string;
  status: BuildTimelineItemStatus;
}