# lib/api-zod/src/generated/types/builderConnection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure controller API for an Android Termux ZIP-to-APK builder
 * OpenAPI spec version: 1.0.0
 */
import type { BuilderConnectionEnvironment } from './builderConnectionEnvironment';
import type { BuilderConnectionStatus } from './builderConnectionStatus';
import type { BuilderConnectionToolsItem } from './builderConnectionToolsItem';

export interface BuilderConnection {
  connected: boolean;
  status: BuilderConnectionStatus;
  tools: BuilderConnectionToolsItem[];
  checkedAt: Date;
  environment?: BuilderConnectionEnvironment;
}